move pokemon lookup command to new pattern. Custom missingno code currently causes a crash, but that's what you get for looking up missingno I guess
